meta-charset/meta-inside-title.html

<!DOCTYPE html>
<title><meta charset=windows-1251></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => assert_equals(document.charset, 'windows-1251'), "charset inside title should work");
</script>
